Final Polish done, README.md updated

+++ main/ttl_checker.py
@@ -6,25 +6,19 @@ from django.utils.timezone import utc
 
 MAX_TIME_OUT = 5.0
 
-
-def get_time_diff(data):
-    if data.updated_at:
-        now = datetime.utcnow().replace(tzinfo=utc)
-        timediff = now - data.updated_at
-        return timediff.total_seconds()/60
-
 def background_process():
+    print('Background TTL process running...')
+    
     while True:
         dataset = Data.objects.all()
-
-        for data in dataset:
-            print(get_time_diff(data))
-            if(get_time_diff(data)>=MAX_TIME_OUT):
-                print('{} Deleted'.format(data))
-                data.delete()
-                
+        if not dataset.count() == 0:
+            for data in dataset:
+                print('Time elapsed for {} = {}'.format(data,get_time_diff(data)))
+                if(get_time_diff(data)>=MAX_TIME_OUT):
+                    print('{} Deleted'.format(data))
+                    data.delete()
+            print('--------------------------------------------------')
         time.sleep(5)
-        print('Background iteration completed')
 
 
 
@@ -37,16 +31,9 @@ def update_all_ttl():
     Data.objects.all().update(updated_at = datetime.utcnow().replace(tzinfo=utc))
 
 
+def get_time_diff(data):
+    if data.updated_at:
+        now = datetime.utcnow().replace(tzinfo=utc)
+        timediff = now - data.updated_at
+        return timediff.total_seconds()/60
 
-
-
-''' Demo Code to run it 
-def task():
-    import threading
-    t = threading.Thread(target=background_process, args=(), kwargs={})
-    t.setDaemon(True)
-    t.start()
-task()
-
-
-'''
